Refactor import statements in Utils.ts and test files
